backend/requirements.txt
# Le-Route Backend Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Mistral AI
mistralai>=1.0.0

# Data processing
numpy>=1.26.0
scikit-learn>=1.4.0

# ML/Router
torch>=2.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# HTTP client (for async requests)
httpx>=0.26.0

# Progress bars
tqdm>=4.66.0
